requirements.txt
asttokens==2.0.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
executing==0.5.4
icecream==2.1.0
idna==2.10
mysql-connector-python==8.0.23
protobuf==3.15.3
Pygments==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
